Corrected DDD structure
