+++ server/bookings/mutations/deleteBooking.js
@@ -1,4 +1,4 @@
-const Booking = require("../bookingModel");
+const Booking = require("../../models/bookingModel");
 
 async function deleteBooking(req, res) {
   try {
