+++ server/bookings/queries/getAllBookings.js
@@ -1,4 +1,4 @@
-const Booking = require("../bookingModel");
+const Booking = require("../../models/bookingModel");
 
 async function getAllBookings(req, res) {
   try {
