+++ server/bookings/queries/getBookingsByUserEmail.js
@@ -1,4 +1,4 @@
-const Booking = require("../bookingModel");
+const Booking = require("../../models/bookingModel");
 
 async function getBookingsByUserEmail(req, res) {
   try {
