+++ server/businesses/mutations/updateBusiness.js
@@ -1,5 +1,5 @@
-const { businessSchema } = require("../validate");
-const Business = require("../businessModel");
+const { businessSchema } = require("../../utils/validateBusiness");
+const Business = require("../../models/businessModel");
 
 async function updateBusiness(req, res) {
   const { error } = businessSchema.validate(req.body);
