+++ server/utils/generateToken.js
